rusha_django_project/pyproject.toml
[tool.poetry]
name = "rusha_django"
version = "0.1.0"
description = ""
authors = ["Emmanuel <emmanuel.davidson@thejitu.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
PyYAML = "^6.0"
ansible = "^6.5.0"
redis = "^4.4.0"
django-redis = "^5.2.0"
celery = "^5.2.7"
psycopg2 = "^2.9.5"
jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
